Only pull changes and re-build website when the master updated

+++ deploy.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
+
+# Intended to be called as a cron to deploy from git master
 cd "$(dirname "$0")"
 git fetch --all
-git reset --hard origin/master
-/usr/bin/hugo -D -d public/
+if [[ $(git status --porcelain) ]]; then
+    git reset --hard origin/master
+    /usr/bin/hugo -D -d public/
+    echo "$(date +\"%Y-%m-%d_%H-%M-%S\") Rebuilt website" > /var/log/"$0"
+fi
